% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{CheBall}
\alias{CheBall}
\title{Compute the Chebychev ball of a H-polytope, P:= Ax<=b}
\usage{
CheBall(A, b)
}
\arguments{
\item{A}{the matrix of the H-polytope}

\item{b}{the vector with the constants of the hyperplanes}
}
\value{
The Chebychev center of the Polytope discribed by the matrix \code{A} and the vector \code{b}
}
\description{
Compute the Chebychev ball of a H-polytope, P:= Ax<=b
}
\examples{
CheBall(A,b)
}
