% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{copula}
\alias{copula}
\title{Construct a copula using uniform sampling from the unit simplex}
\usage{
copula(R1 = NULL, R2 = NULL, Sigma = NULL, M = NULL, N = NULL)
}
\arguments{
\item{M}{The number of the slices for the copula. The default value is 100.}

\item{N}{The number of points to sample. The default value is \eqn{5\cdot 10^5}.}

\item{h1}{A \eqn{d}-dimensional vector that describes the direction of the first family of parallel hyperplanes.}

\item{h2}{Optional. A \eqn{d}-dimensional vector that describes the direction of the second family of parallel hyperplanes.}

\item{E}{Optional. The \eqn{d\times d} symmetric positive semidefine matrix that describes the family of concentric ellipsoids centered at the origin.}
}
\value{
A \eqn{numSlices\times numSlices} numerical matrix that corresponds to a copula.
}
\description{
Given two families of parallel hyperplanes or a family of parallel hyperplanes and a family of concentric ellispoids centered at the origin intersecting the canonical simplex, this function uniformly samples from the canonical simplex and construct an approximation of the bivariate probability distribution, called copula.
}
\examples{
# compute a copula for two random families of parallel hyperplanes
h1 = runif(n = 10, min = 1, max = 1000)
h1 = h1 / 1000
h2=runif(n = 10, min = 1, max = 1000)
h2 = h2 / 1000
cop = copula(R1 = h1, R2 = h2, M = 10, N = 100000)

# compute a copula for a family of parallel hyperplanes and a family of conentric ellipsoids
h = runif(n = 10, min = 1, max = 1000)
h = h / 1000
E = replicate(10, rnorm(20))
E = cov(E)
cop = copula(R1 = h, Sigma = E, M = 10, N = 100000)

}
\references{
\cite{L. Cales, A. Chalkis, I.Z. Emiris, V. Fisikopoulos,
\dQuote{Practical volume computation of structured convex bodies, and an application to modeling portfolio dependencies and financial crises,} \emph{Proc. of Symposium on Computational Geometry, Budapest, Hungary,} 2018.}
}
