% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoSampling.R
\name{demoSampling}
\alias{demoSampling}
\title{Run some sampling experiments.}
\usage{
demoSampling(distribution)
}
\arguments{
\item{uniform}{The string "uniform" to choose uniform as the target distribution.}

\item{gaussian}{The string "gaussian" to choose spherical gaussian as the target distribution.}
}
\value{
Print the computed volumes and the error. If the test fails a message is printed.
}
\description{
Use uniform or spherical gaussian to sample from some convex H-polytopes, i.e. cubes, simplices, skinny_cubes, cross polytopes, birkhoff polytopes.
We use the default values, i.e. \eqn{walk length = \lfloor 10+dimension/10\rfloor}, \eqn{N = 100}, Cordinate Directions HnR, \eqn{variance = 1}.
}
\examples{
#choose uniform distribution
demoSampling("uniform")
#choose spherical gaussian distribution
demoSampling("gaussian")
}
