% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoVolEsti.R
\name{demoVolEsti}
\alias{demoVolEsti}
\title{Run some experiments}
\usage{
demoVolEsti()
}
\value{
Print the computed volumes and the total time
}
\description{
Run some experiments
}
\examples{
testRvolEsti()
}
