% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoVolume.R
\name{demoVolume}
\alias{demoVolume}
\title{Run some volume approxiamtion experiments.}
\usage{
demoVolume(algo)
}
\arguments{
\item{CV}{The string "CV" to choose CV algorithm.}

\item{volesti}{The string "volesti" to choose volesti algorithm.}
}
\value{
Print the computed volumes and the error. If the test fails a message is printed.
}
\description{
Run volesti or CV algorithm to approximate the volume of some cubes, simplices, skinny_cubes, cross polytopes, birkhoff polytopes.
We run \eqn{10} experiments for volesti and \eqn{20} for CV. We demand \eqn{error = 0.1}. For all the other parameters use the default values for both algorithms.
}
\examples{
#test volesti
demoVolume("volesti")
#test CV
demoVolume("CV")
}
