% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lipschitz_constant.R
\name{dinvweibull_with_loc}
\alias{dinvweibull_with_loc}
\title{Inverse weibull distribution PDF with location parameter}
\usage{
dinvweibull_with_loc(x, k, lambda, theta)
}
\description{
Inverse weibull distribution PDF with location parameter
}
