% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ess}
\alias{ess}
\title{Gelman-Rubin and Brooks-Gelman Potential Scale Reduction Factor (PSRF) for each marginal}
\usage{
ess(samples)
}
\arguments{
\item{samples}{A matrix that contans column-wise the sampled points from a geometric random walk.}
}
\value{
A vector that contains the values of PSRF for each coordinate
}
\description{
Gelman-Rubin and Brooks-Gelman Potential Scale Reduction Factor (PSRF) for each marginal
}
\references{
\cite{Gelman, A. and Rubin, D. B.,
\dQuote{Inference from iterative simulation using multiple sequences,} \emph{Statistical Science,} 1992.}

\cite{Brooks, S. and Gelman, A.,
\dQuote{General Methods for Monitoring Convergence of Iterative Simulations,} \emph{Journal of Computational and Graphical Statistics,} 1998.}
}
