% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lipschitz_constant.R
\name{estimtate_lipschitz_constant}
\alias{estimtate_lipschitz_constant}
\title{Estimate the Lipschitz Constant of a function f}
\usage{
estimtate_lipschitz_constant(f, P, n, m = 1)
}
\description{
Estimate the Lipschitz Constant of a function f
}
