% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactZonoVol.R
\name{ExactZonoVol}
\alias{ExactZonoVol}
\title{Compute the exact volume of a zonotope}
\usage{
ExactZonoVol(ZonoMat)
}
\arguments{
\item{ZonoMat}{The \eqn{m \times d} matrix that containes the segments that define the zonotope.}
}
\value{
The exact volume of the zonotope
}
\description{
Given the \eqn{m \times d} matrix that containes the \eqn{m} segments that define the \eqn{d}-dimensional zonotope, this function computes the sum of the absolute values of the determinants of all the \eqn{d \times d} submatrices.
}
\examples{

# compute the exact volume of a 5-dimensional zonotope defined by the Minkowski sum of 10 segments
ZonoMat = GenZonotope(5, 10)
vol = ExactZonoVol(ZonoMat)
}
