% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{full_dimensional_polytope}
\alias{full_dimensional_polytope}
\title{Internal rcpp function to compute the full dimensional polytope when a low dimensional is given}
\usage{
full_dimensional_polytope(P)
}
\arguments{
\item{P}{A low dimensional convex polytope in H-representation.}
}
\value{
A numerical matrix that describes the full dimensional polytope, a numerical matrix of the inverse
        linear transformation that is applied on the input polytope, the numerical vector - point that the
        input polytope is shifted and the product of the singular values of the matrix of the linear map 
        applied on the input polytope.
}
\description{
Internal rcpp function to compute the full dimensional polytope when a low dimensional is given
}
\keyword{internal}
