% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_birkhoff.R
\name{gen_birkhoff}
\alias{gen_birkhoff}
\title{Generator function for Birkhoff polytope}
\usage{
gen_birkhoff(n)
}
\arguments{
\item{n}{The order of the Birkhoff polytope}
}
\value{
A polytope class representing the full dimensional \eqn{n}-Birkhoff polytope in H-representation.
}
\description{
This function can be used to generate the full dimensional \eqn{n}-Birkhoff polytope in H-representation.
}
\examples{
# generate the Birkhoff polytope of order 5
P = gen_birkhoff(5)
}
