% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_prod_simplex.R
\name{gen_prod_simplex}
\alias{gen_prod_simplex}
\title{Generator function for product of simplices}
\usage{
gen_prod_simplex(dimension)
}
\arguments{
\item{dimension}{The dimension of the simplices.}
}
\value{
A polytope class representing the product of the two \eqn{d}-dimensional unit simplices in H-representation.
}
\description{
This function generates a \eqn{2d}-dimensional polytope that is defined as the product of two \eqn{d}-dimensional unit simplices in H-representation.
}
\examples{
# generate a product of two 5-dimensional simplices.
P = gen_prod_simplex(5)
}
