% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_rand_vpoly.R
\name{gen_rand_vpoly}
\alias{gen_rand_vpoly}
\title{Generator function for random V-polytopes}
\usage{
gen_rand_vpoly(dimension, m)
}
\arguments{
\item{dimension}{The dimension of the convex polytope.}

\item{m}{The number of the vertices.}
}
\value{
A polytope class representing a V-polytope.
}
\description{
This function can be used to generate a \eqn{d}-dimensional polytope in V-representation with \eqn{m} vertices. We pick \eqn{m} random points from the boundary of the \eqn{d}-dimensional unit hypersphere as vertices.
}
\examples{
# generate a 10-dimensional polytope defined as the convex hull of 25 random vertices
P = gen_rand_vpoly(10, 25)
}
