% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenCube.R
\name{GenCube}
\alias{GenCube}
\title{Generator function for hypercubes.}
\usage{
GenCube(DimGen, repr)
}
\arguments{
\item{Dimension}{The dimension of the hypercube}

\item{represenation}{A string to declare the representation. It has to be H for H-representation or V for V-representation.}
}
\value{
A hypercube in H or V-representation. For an H polytope the return value is a list with two elements: the "matrix" containing a \eqn{2d\times d} matrix A and the "vector" containing a 2d-dimensional vector b, s.t. \eqn{Ax\leq b}. When the V-representation is chosen the return value is a \eqn{2^d \times d} matrix that containes the vertices row-wise.
}
\description{
This function can be used to generate a d-dimensional Hypercube \eqn{[-1,1]^d} in H or V representation.
}
\examples{
# generate a 10-dimensional hypercube in H-representation
PolyList = GenCube(10, 'H')

# generate a 15-dimension hypercube in V-representation
PolyList = GenCube(15, 'V')
}
