% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenRandHpoly.R
\name{GenRandHpoly}
\alias{GenRandHpoly}
\title{Generator function for random H-polytopes}
\usage{
GenRandHpoly(dimension, m)
}
\arguments{
\item{dimension}{The dimension of the convex polytope.}

\item{m}{The number of the facets.}
}
\value{
A polytope class representing a H-polytope.
}
\description{
This function can be used to generate a \eqn{d}-dimensional polytope in H-representation with \eqn{m} facets. We pick \eqn{m} random hyperplanes tangent on the \eqn{d}-dimensional unit hypersphere as facets.
}
\examples{
# generate a 10-dimensional polytope with 50 facets
P = GenRandVpoly(10, 50)
}
