% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSkinnyCube.R
\name{GenSkinnyCube}
\alias{GenSkinnyCube}
\title{Generator function for skinny hypercubes.}
\usage{
GenSkinnyCube(DimGen, repr = "H")
}
\arguments{
\item{Dimension}{The dimension of the skinny hypercube.}
}
\value{
A d-dimensional skinny hypercube in H-representation.The retutn value is a list with two elements: the "matrix" containing a \eqn{2d\times d} matrix A and the "vector" containing a 2d-dimensional vector b, s.t. \eqn{Ax\leq b}.
}
\description{
This function can be used to generate a d-dimensional skinny hypercube.
}
\examples{
# generate a 10-dimensional skinny hypercube.
PolyList = GenSkinnyCube(10)
}
