% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_full_dimensional_polytope.R
\name{get_full_dimensional_polytope}
\alias{get_full_dimensional_polytope}
\title{Given a low-dimensional H-polytope, restrict it to its null space}
\usage{
get_full_dimensional_polytope(P)
}
\arguments{
\item{P}{A low-dimensional convex polytope in H-representation.}
}
\value{
A list with 4 elements: (a) a full-dimensional polytope in H-representation and (b) the element "N" which is the matrix of the inverse linear transformation that is applied on the input polytope, (c)  the element "shift" which is the opposite vector of that which has shifted the input polytope, (d) the element "svd_prod" which is the product of the singular values of the matrix N and it can be used for further volume estimations.
}
\description{
Given a low-dimensional H-polytope, restrict it to its null space
}
\examples{
P = gen_cube(10,'H')
P$Aeq = gen_rand_hpoly(10,3)$A
P$beq = rep(0,3)
ret_list = get_full_dimensional_polytope(P)
}
