% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geweke}
\alias{geweke}
\title{Geweke MCMC diagnostic test}
\usage{
geweke(samples, frac1 = NULL, frac2 = NULL)
}
\arguments{
\item{frac1}{Optional. A portion of the first in order points in matrix samples.}

\item{frac2}{Optional. A portion of the last in order points in matrix samples.}

\item{A}{matrix that contans column-wise the sampled points from a geometric random walk.}
}
\value{
A boolean to denote if the test is passes or fails
}
\description{
Geweke MCMC diagnostic test
}
\references{
\cite{Geweke, J.,
\dQuote{Evaluating the accuracy of sampling-based approaches to the calculation of posterior moments,} \emph{ In Bayesian Statistics 4. Proceedings of the Fourth Valencia International Meeting,} 1992.}
}
