% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HdemoVolume.R
\name{HdemoVolume}
\alias{HdemoVolume}
\title{Run some volume approximation experiments for H-polytopes.}
\usage{
HdemoVolume(algo)
}
\arguments{
\item{CG}{The string "CG" to choose CoolingGaussian algorithm.}

\item{SOB}{The string "SOB" to choose SequenceOfBalls algorithm.}
}
\value{
Print the computed volumes and the error. If the test fails a message is printed.
}
\description{
Choose between SequenceOfBalls and CoolingGaussian algorithm to approximate the volume of some cubes, simplices, skinny_cubes, cross polytopes and birkhoff polytopes in H-representation.
For each polytope we run \eqn{10} volume experiments for SequenceOfBalls and \eqn{20} for CoolingGaussian and we consider the mean value as the volume approximation. We demand \eqn{error = 0.1} for the most of them. For all the other parameters we use the default values for both algorithms.
}
\examples{
# test SequenceOfBalls
HdemoVolume("SOB")
# test CoolingGausian
HdemoVolume("CG")
}
