% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytopes.R
\docType{class}
\name{HPolytope-class}
\alias{HPolytope-class}
\alias{HPolytope}
\title{A reference class to represent a H-polytope}
\description{
A H-polytope is a convex polytope defined by a set of linear inequalities or equivalently a \eqn{d}-dimensional H-polytope with \eqn{m} facets is defined by a \eqn{m\times d} matrix A and a \eqn{m}-dimensional vector b, s.t.: \eqn{Ax\leq b}.
}
\section{Fields}{

\describe{
\item{\code{A}}{\eqn{m\times d} numerical matrix A}

\item{\code{A}}{\eqn{m}-dimensional vector b}
}}

\section{Methods}{

\describe{
\item{\code{get_mat()}}{Return a numerical matrix that describes the polytope in ine format}
}}

