% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadSdpaFormatFile}
\alias{loadSdpaFormatFile}
\title{Read a SDPA format file}
\usage{
loadSdpaFormatFile(inputFile = NULL)
}
\value{
A list with two named items: an item "matrices" which is a list of the matrices and an vector "objFunction"
}
\description{
Read a SDPA format file
}
\section{Fields}{

\describe{
\item{\code{inputFile}}{Name of the input file}
}}

\examples{
l = loadSdpaFormatFile("input.txt")
}
