% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{ode_solve}
\alias{ode_solve}
\title{Solve an ODE of the form dx^n / dt^n = F(x, t)}
\usage{
ode_solve(
  n,
  step_size,
  order,
  dimension,
  initial_time,
  F,
  method,
  domains = NULL,
  initial_conditions = NULL
)
}
\arguments{
\item{n}{The number of steps.}

\item{step_size}{The step size.}

\item{order}{The ODE order (default is n = 1)}

\item{dimension}{The dimension of each derivative}

\item{initial_time}{The initial time}

\item{F}{The function oracle F(x, t) in the ODE.}

\item{method}{The method to be used}

\item{domains}{A list of n H-polytopes with keys "P_1", "P_2", ..., "P_n" that correspond to each derivative's domain}

\item{initial_conditions}{The initial conditions provided to the solver. Must be provided in a list with keys "x_1", ..., "x_n" and column vectors as values. The state "x_n" represents the (n-1)-th order derivative with respect to time}
}
\value{
A list which contains elements "x_1", ..., "x_n" representing each derivative results. Each "x_i" corresponds to a d x n matrix where each column represents a certain timestep of the solver.
}
\description{
Solve an ODE of the form dx^n / dt^n = F(x, t)
}
\examples{
# Please visit the examples directory on examples demonstrating usage of the ODE solvers.

}
