% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psrf}
\alias{psrf}
\title{Gelman-Rubin Potential Scale Reduction Factor (PSRF)}
\usage{
psrf(samples)
}
\arguments{
\item{samples}{A matrix that contans column-wise the sampled points from a geometric random walk.}
}
\value{
The value of PSRF diagnostic.
}
\description{
Gelman-Rubin Potential Scale Reduction Factor (PSRF)
}
\references{
\cite{Gelman, A. and Rubin, D. B.,
\dQuote{Inference from iterative simulation using multiple sequences,} \emph{Statistical Science,} 1992.}
}
