% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{psrf_multivariate}
\alias{psrf_multivariate}
\title{Gelman-Rubin Potential Scale Reduction Factor (PSRF)}
\usage{
psrf_multivariate(samples)
}
\arguments{
\item{samples}{A matrix that contans column-wise the sampled points from a geometric random walk.}
}
\value{
The value of multivariate PSRF by S. Brooks and A. Gelman.
}
\description{
Gelman-Rubin Potential Scale Reduction Factor (PSRF)
}
\references{
\cite{Gelman, A. and Rubin, D. B.,
\dQuote{Inference from iterative simulation using multiple sequences,} \emph{Statistical Science,} 1992.}

\cite{Brooks, S. and Gelman, A.,
\dQuote{General Methods for Monitoring Convergence of Iterative Simulations,} \emph{Journal of Computational and Graphical Statistics,} 1998.}
}
