% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{raftery}
\alias{raftery}
\title{Raftery and Lewis MCMC diagnostic}
\usage{
raftery(samples, q = NULL, r = NULL, s = NULL)
}
\arguments{
\item{samples}{A matrix that contans column-wise the sampled points from a geometric random walk.}

\item{q}{Optional. The quantile of the quantity of interest. The default value is 0.025.}

\item{r}{Optional. The level of precision desired. The default value is 0.01.}

\item{s}{Optional. The probability associated with r. The default value is 0.95.}
}
\value{
(i) The number of draws required for burn-in, (ii) the skip parameter for 1st-order Markov chain, (iii) the skip parameter sufficient to get independence chain, (iv) the number of draws required to achieve r precision, (v) the number of draws if the chain is white noise, (vi) the I-statistic from Raftery and Lewis (1992).
}
\description{
Raftery and Lewis MCMC diagnostic
}
\references{
\cite{Raftery, A. E. and Lewis, S. M.,
\dQuote{How many iterations in the Gibbs sampler?,} \emph{Bayesian Statistics 4. Proceedings of the Fourth Valencia International Meeting,} 1992.}
}
