% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{randomized_lp_solver}
\alias{randomized_lp_solver}
\title{Solving a LP problem using Randomized algorithms.}
\usage{
randomized_lp_solver(P, obj, bounds = NULL, algo = 0L, verbose = FALSE)
}
\arguments{
\item{bounds}{optional. A list that contains the bound of the variables (the default is complete Real space), as follows:
\itemize{
\item{\code{indices}}{A vector containing the variable indices(0 indexed) for which the bounds have to be set.}
\item{\code{lower}}{A vector containing the value of lower bounds for all the variables specified in indices.}
\item{\code{upper}}{A vector containing the value of upper bounds for all the variables specified in indices.}
}}

\item{algo}{Optional. An unsigned integer that declares which algorithm, as follows:
\itemize{
\item{\code{0} }{ Use the Randomized Cutting Plane algorithm (RCP).}
\item{\code{1} }{ Use the Simulated Annealing algorithm (SIM_ANN).}
}}

\item{verbose}{Optional. A boolean parameter for printing out the LP program formed.}

\item{P.}{A convex H Polytope, it is the feasible region of the LP problem (\eqn{Ax<=b}).}

\item{obj.}{A vector for the coefficients of the objective function (\eqn{min c^{T} x}).}
}
\value{
A list containing the value of the objective function and value of all variables.
}
\description{
Currently, only 2 algorithms are implemented - Randomized Cutting plane and Simulated Annealing.
}
\examples{
# computing Chebychev ball for a H-polytope (3d cube)
P <- gen_cube(3, 'H')
row_norm <- sqrt(rowSums((P$A)^2))
P$A <- cbind(P$A, row_norm)
var_bounds <- list("indices"=c(3), "lower"=c(0), "upper"=c(1000))
randomized_lp_solver(P, obj=c(0,0,0,-1), bounds=var_bounds, algo=1, verbose=TRUE)
}
\references{
\cite{ Dabbene, Fabrizio, Pavel S. Shcherbakov, and Boris T. Polyak.,
\dQuote{ A randomized cutting plane method with probabilistic geometric convergence,} \emph{SIAM Journal on Optimization 20.6,} (2010): 3185-3207.},

\cite{Adam Tauman Kalai, Santosh Vempala,
\dQuote{Simulated Annealing for Convex Optimization,} \emph{Mathematics of Operations Research Vol. 31, No. 2,} 2006.}
}
