\docType{class}
\name{Rcpp_Hpolytope}
\alias{Rcpp_Hpolytope-class}
\alias{[,Rcpp_Hpolytope-method}
\alias{[,Rcpp_Hpolytope,ANY,ANY,ANY-method}
\alias{$<-,Rcpp_Hpolytope-method}
\alias{$,Rcpp_Hpolytope-method}
\alias{filepaths<-,Rcpp_Hpolytope-method}
\title{
Rcpp_Hpolytopes
}
\description{
this is Hpoly
}
\details{
\describe{
   \item{type}{An integer for the representation.}

\item{dimension}{An integer for the dimension.}

\item{A}{A numerical matrix.}

\item{b}{A numerical vector.}
 }
}

