\docType{class}
\name{Rcpp_IntVP}
\alias{Rcpp_IntVP-class}
\alias{[,Rcpp_IntVP-method}
\alias{[,Rcpp_IntVP,ANY,ANY,ANY-method}
\alias{$<-,Rcpp_IntVP-method}
\alias{$,Rcpp_IntVP-method}
\alias{filepaths<-,Rcpp_IntVP-method}
\title{
Rcpp_IntVP
}
\description{
this is IntVP
}
\details{
\describe{
   \item{type}{An integer for the representation.}

\item{dimension}{An integer for the dimension.}

\item{V1}{A numerical matrix.}

\item{V2}{A numerical matrix.}
 }
}
