\docType{class}
\name{Rcpp_Vpolytope}
\alias{Rcpp_Vpolytope-class}
\alias{[,Rcpp_Vpolytope-method}
\alias{[,Rcpp_Vpolytope,ANY,ANY,ANY-method}
\alias{$<-,Rcpp_Vpolytope-method}
\alias{$,Rcpp_Vpolytope-method}
\alias{filepaths<-,Rcpp_Vpolytope-method}
\title{
Rcpp_Vpolytopes
}
\description{
this is Vpoly
}
\details{
\describe{
   \item{type}{An integer for the representation.}

\item{dimension}{An integer for the dimension.}

\item{V}{A numerical matrix.}

 }
}
