\docType{class}
\name{Rcpp_Zonotope}
\alias{Rcpp_Zonotope-class}
\alias{[,Rcpp_Zonotope-method}
\alias{[,Rcpp_Zonotope,ANY,ANY,ANY-method}
\alias{$<-,Rcpp_Zonotope-method}
\alias{$,Rcpp_Zonotope-method}
\alias{filepaths<-,Rcpp_Zonotope-method}
\title{
Rcpp_Zonotope
}
\description{
this is Zonotope
}
\details{
\describe{
   \item{type}{An integer for the representation.}

\item{dimension}{An integer for the dimension.}

\item{G}{A numerical matrix.}
 }
}
