% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rotating}
\alias{rotating}
\title{An internal Rccp function for the random rotation of a convex polytope}
\usage{
rotating(P)
}
\arguments{
\item{P}{A convex polytope (H-, V-polytope or a zonotope).}
}
\value{
A matrix that describes the rotated polytope
}
\description{
An internal Rccp function for the random rotation of a convex polytope
}
\section{warning}{

Do not use this function.
}

