% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rounding}
\alias{rounding}
\title{Internal rcpp function for the rounding of a convex polytope.}
\usage{
rounding(P, WalkType = NULL, walk_length = NULL, radius = NULL)
}
\arguments{
\item{WalkType}{Optional. A string that declares the random walk.}

\item{walk_length}{Optional. The number of the steps for the random walk.}

\item{radius}{Optional. The radius for the ball walk.}
}
\value{
A Matrix that describes the rounded polytope and contains the round value.
}
\description{
Internal rcpp function for the rounding of a convex polytope.
}
