% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sample_points}
\alias{sample_points}
\title{Sample many points from a convex Polytope (H-polytope, V-polytope or a zonotope) or use direct methods for uniform sampling from unit simplex and hypersphere}
\usage{
sample_points(P = NULL, N = NULL, WalkType = NULL, walk_len = NULL,
  exact = NULL, body = NULL, Parameters = NULL,
  distribution = NULL, InnerPoint = NULL)
}
\arguments{
\item{P}{A convex polytope. It is an object from class (a) HPolytope or (b) VPolytope or (c) Zonotope.}

\item{N}{The number of points that the function is going to sample from the convex polytope. Default value is \eqn{100}.}

\item{distribution}{Optional. A list that contains parameters for the target distribution. Default distribution is uniform.
\itemize{
 \item{gaussian }{A boolean parameter. It declares spherical gaussian distribution as the target distribution. Default value is false.}
 \item{variance }{The variance for the spherical gaussian distribution. Default value is \eqn{1}.}
}}

\item{InnerPoint}{A \eqn{d}-dimensional numerical vector that defines a point in the interior of polytope P.}

\item{method}{Optional. A list that contains parameters for the random walk method. Default method is Coordinate Hit-and-Run.
\itemize{
 \item{direct }{A boolean parameter. It should be used for uniform sampling from the boundary or the interior of a hypersphere or from a unit or an arbitrary simplex. The arbitrary simplex has to be given as a V-polytope and the dimension should not be declared through method list. For the rest well known convex bodies it has to be declared the dimension and the type of body (simplex, sphere, ball).}
 \item{dim }{An integer that declares the dimension when direct flag is enabled for a unit simplex or a hypersphere (boundary or interior).}
 \item{body }{A string to request uniform sampling: (a) "simplex" to sample from an arbitrary simplex (when the simplex is given as a V-polytope) or a unit simplex (when no polytope is given and the dimension is declared), (b) "sphere" to sample from the boundary of a {d}-dimensional hypersphere centered at the origin and (c) to sample from the interior of the \eqn{d}-dimensional hypersphere centered at the origin. For (b) and (c) dimension should be given as well through method list.}
 \item{radius }{The radius of the \eqn{d}-dimensional hypersphere. Default value is \eqn{1}.}
 \item{WalkT }{A string to declare the random walk method: (a)"hnr" for Hit-and-Run or (b) "bw" for ball walk. Default method is Hit-and-Run.}
 \item{coord }{A boolean parameter for the hit-and-run. True for Coordinate Directions HnR, false for Random Directions HnR. Default value is TRUE.}
 \item{delta }{Optional. The radius for the ball walk.}
 \item{W }{Optional. The number of the steps for the random walk. Default value is \eqn{\lfloor 10+d/10\rfloor}.}
}}
}
\value{
A \eqn{d\times N} matrix that contains, column-wise, the sampled points from the convex polytope.
}
\description{
Sample N points from a H or a V-polytope or a zonotope with uniform or spherical gaussian -centered in an internal point- target distribution.
The \eqn{d}-dimensional unit simplex is the set of points \eqn{\vec{x}\in \R^d}, s.t.: \eqn{\sum_i x_i\leq 1}, \eqn{x_i\geq 0}. The \eqn{d}-dimensional canonical simplex is the set of points \eqn{\vec{x}\in \R^d}, s.t.: \eqn{\sum_i x_i = 1}, \eqn{x_i\geq 0}.
}
\examples{
# uniform distribution from a 3d cube in V-representation using ball walk
P = GenCube(3, 'V')
points = sample_points(P, method = list("WalkT"="bw", "W"=5))

# gaussian distribution from a 2d unit simplex in H-representation with variance = 2
A = matrix(c(-1,0,0,-1,1,1), ncol=2, nrow=3, byrow=TRUE)
b = c(0,0,1)
P = HPolytope(A=A, b=b)
points = sample_points(P, distribution = list("gaussian"=TRUE, "variance"=2))
}
\references{
\cite{R.Y. Rubinstein and B. Melamed,
\dQuote{Modern simulation and modeling} \emph{ Wiley Series in Probability and Statistics,} 1998.}

\cite{A Smith, Noah and W Tromble, Roy,
\dQuote{Sampling Uniformly from the Unit Simplex,} \emph{ Center for Language and Speech Processing Johns Hopkins University,} 2004.}

\cite{Art B. Owen,
\dQuote{Monte Carlo theory, methods and examples,} \emph{ Copyright Art Owen,} 2009-2013.}
}
