% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sphere.R
\name{sample_sphere}
\alias{sample_sphere}
\title{Function to sample uniformly from a sphere or a ball}
\usage{
sample_sphere(dimension, radius, N, interior)
}
\arguments{
\item{dimension}{The dimension of the sphere.}

\item{radius}{The radius of the sphere.}

\item{N}{The number of the points to sample.}

\item{interior}{A boolean flag. It has to be TRUE when sampling is from the interior of the sphere.}
}
\value{
A \eqn{d\times N} matrix that containes, column-wise, the sampled points.
}
\description{
This function can be used for uniform sampling from a \eqn{d}-dimensional sphere of radius \eqn{r} or from the interior of the sphere (ball).
}
\examples{
# sample 1200 random points from a 10-dimensional hypersphere of radius 100
points = sample_sphere(dimension = 10, radius = 100, N = 1200)

# sample 1000 random points from a 20-dimensional ball of radius 200
points = sample_sphere(dimension = 20, radius = 200, N = 1000, interior = TRUE)
}
