% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{testRvolEsti}
\alias{testRvolEsti}
\title{Run some experiments}
\usage{
testRvolEsti()
}
\value{
Print the computed volumes and the total time
}
\description{
Run some experiments
}
\examples{
testRvolEsti()
}
