% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VdemoSampling.R
\name{VdemoVolume}
\alias{VdemoVolume}
\title{Run some volume approximation experiments for V-polytopes.}
\usage{
VdemoVolume(algo)
}
\arguments{
\item{CG}{The string "CG" to choose CoolingGaussian algorithm.}

\item{SOB}{The string "SOB" to choose SequenceOfBalls algorithm.}
}
\value{
Print the computed volumes and the error. If the test fails a message is printed.
}
\description{
Run SequenceOfBalls or CoolingGaussian algorithm to approximate the volume of some cubes, simplices and cross polytopes in V-representation.
For each polytope we run \eqn{10} experiments.
}
\examples{
# test SequenceOfBalls
VdemoVolume("SOB")
# test CoolingGausian
VdemoVolume("CG")
}
