% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{vol_R}
\alias{vol_R}
\title{The Cpp main function compiled by Rcpp and used by the main R function VolEsti for volume approximation}
\usage{
vol_R(A, W, e, Chebychev, coord, rounding, V)
}
\arguments{
\item{A}{the matrix of the H-polytope in ine format}

\item{W}{The number of steps of the random walk}

\item{e}{The goal for the error of the approximation}

\item{Chebychev}{A d+1 numerical vector desribing the Chebychev ball. If it is of length d+2 then it is considered as null and lpSolve library is used for the computation of the Chebychev ball}

\item{coord}{A boolean parameter. True for Coordinate Directions HnR, false for Random Directions HnR}

\item{rounding}{A boolean parameter. True to activate rounding option, false to deactivate}

\item{V}{A boolean parameter. True to activate printing, false to deactivate}
}
\value{
An approximation of the Polytope described by matrix \code{A} in ine format
}
\description{
The Cpp main function compiled by Rcpp and used by the main R function VolEsti for volume approximation
}
\examples{
Vol_R(A,10,0.5,c(0,0,0,0,0),TRUE,FALSE,FALSE)
}
