% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{VolEsti}
\alias{VolEsti}
\title{The main R function for volume approximation of a convex H-Polytope}
\usage{
VolEsti(Inputs)
}
\arguments{
\item{list("path", "matrix", "vector", "Chebychev", "verbose", "coordinate", "rounding", "Walk_length", "error", "test")}{A list that includes alla the parameters of the algorithm}

\item{path}{The path to the ine file that describes the H-polytope. If path is given then "matrix" and "vector" inputs are not needed}

\item{matrix}{The matrix A of the polytope. If it is in ine format then the input "vector" is not needed}

\item{vector}{The vector b that containes the constants of the hyperplanes}

\item{Chebychev}{Optional. A d+1 vector that containes the chebychev center in the first d coordinates and the radius of the chebychev ball in the last coordinate}

\item{verbose}{Optional. A boolean parameter for printing. Default is False}

\item{coordinate}{Optional. A boolean parameter for the hit-and-run. True for Coordinate Directions HnR, false for Random Directions HnR. Default is True}

\item{rounding}{Optional. A boolean parameter to activate the rounding option. Default is False}

\item{Walk_length}{Optional. Declare the number of the steps for the random walk, default is 10+d/10}

\item{error}{Optional. Declare the goal for the approximation error. Default is 1}

\item{test}{Optional. A boolean parameter. Declare if the current excecution is a test or not. Default is False}
}
\value{
The approximation of the volume of an H-polytope
}
\description{
The main R function for volume approximation of a convex H-Polytope
}
\examples{
VolEsti(list("path"=/path/to/ine/file, "verbose"=TRUE))
}
