% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytopes.R
\docType{class}
\name{VPolytope-class}
\alias{VPolytope-class}
\alias{VPolytope}
\title{A reference class to represent a V-polytope or the intersection of two V-polytopes}
\description{
A V-polytope is a convex polytope defined by the set of its vertices.
}
\section{Fields}{

\describe{
\item{\code{A}}{\eqn{m\times d} numerical matrix V that contains row-wise the \eqn{m} vertices of a V-polytope.}

\item{\code{A}}{\eqn{k\times d} numerical matrix V2 that contains row-wise the \eqn{k} vertices of the second V-polytope that defines the intersection. If this field is NULL then the created object will represent a V-polytope defined by the first field.}
}}

\section{Methods}{

\describe{
\item{\code{get_mat()}}{Return a numerical matrix that describes the first polytope in ext format}

\item{\code{get_mat2()}}{Return a numerical matrix that describes the second polytope in ine format}
}}

