% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polytopes.R
\docType{class}
\name{Zonotope-class}
\alias{Zonotope-class}
\alias{Zonotope}
\title{A class that represent a zonotope}
\description{
A \eqn{d}-dimensional zonotope is a convex polytope defined by the Minkowski sum of \eqn{d}-dimensional segments.
}
\section{Fields}{

\describe{
\item{\code{A}}{\eqn{m\times d} numerical matrix G that contains row-wise the \eqn{m} segments.}
}}

\section{Methods}{

\describe{
\item{\code{get_mat()}}{Return a numerical matrix that describes the zonotope in ine format}
}}

