% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoRounding.R
\name{demoRounding}
\alias{demoRounding}
\title{Run rounding and rotating tests.}
\usage{
demoRounding(algo)
}
\arguments{
\item{CG}{The string "CG" to choose CoolingGaussian algorithm}

\item{SOB}{The string "SOB" to choose SequenceOfBalls algorithm}
}
\value{
Print the computed volume and print a failure message if the error is larger than the expected.
}
\description{
Choose volume algorithm between CoolingGaussian and SequenceOfBalls and run rounding tests for some skinny cubes. In the first test we apply a random rotation as well before the rounding. We run \eqn{10} volume experiments for SequenceOfBalls and \eqn{20} for CoolingGaussian and we consider the mean value as the volume approximation.
}
\examples{
# run tests for SOB algorithm
demoRounding("SOB")

# run tests for CV algorithm
demoRounding("CG")
}
