% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoRounding.R
\name{demoRounding}
\alias{demoRounding}
\title{Run rounding and rotating tests.}
\usage{
demoRounding(algo)
}
\arguments{
\item{CV}{The string "CV" to choose CV algorithm}

\item{volesti}{The string "volesti" to choose volesti algorithm}
}
\value{
Print the computed volume and print a failure message if the error is larger than the expected.
}
\description{
Choose volume algorithm between CV and volesti and run rounding tests for some skinny cubes. In the first test we apply a random rotation as well before the rounding. For we run 10 experiments for volesti and 20 for CV.
}
\examples{
#run tests for volesti algorithm
demoRounding("volesti")

#run tests for CV algorithm
demoRounding("CV")
}
