% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoVolume.R
\name{demoVolume}
\alias{demoVolume}
\title{Run some experiments}
\usage{
demoVolume(algo)
}
\value{
Print the computed volumes and the total time
}
\description{
Run volesti or CV algorithm to approximate the volume of some cubes, simplices, skinny_cubes, cross polytopes, birkhoff polytopes.
}
\examples{
#test volesti
demoVolume("volesti")
#test CV
demoVolume("CV")
}
