% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demoVolume.R
\name{demoVolume}
\alias{demoVolume}
\title{Run some volume approximation experiments.}
\usage{
demoVolume(algo)
}
\arguments{
\item{CG}{The string "CG" to choose CoolingGaussian algorithm.}

\item{SOB}{The string "SOB" to choose SequenceOfBalls algorithm.}
}
\value{
Print the computed volumes and the error. If the test fails a message is printed.
}
\description{
Run SequenceOfBalls or CoolingGaussian algorithm to approximate the volume of some cubes, simplices, skinny_cubes, cross polytopes, birkhoff polytopes.
We run \eqn{10} experiments for SequenceOfBalls and \eqn{20} for CoolingGaussian, for each polytope and we consider the mean as the computed volume. We demand \eqn{error = 0.1}. For all the other parameters use the default values for both algorithms.
}
\examples{
#test SequenceOfBalls
demoVolume("SOB")
#test CoolingGausian
demoVolume("CG")
}
