% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lipschitz_constant.R
\name{dinvweibull_with_loc}
\alias{dinvweibull_with_loc}
\title{Inverse weibull distribution PDF with location parameter}
\usage{
dinvweibull_with_loc(x, k, lambda, theta)
}
\arguments{
\item{x}{The argument of the PDF}

\item{k}{The shape parameter}

\item{lambda}{The scale parameter}

\item{theta}{The location parameter}
}
\value{
The value of the PDF of an Inverse Weibull distribution with parameters k, lambda, theta evaluated at x
}
\description{
Inverse weibull distribution PDF with location parameter
}
