% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ExactZonoVol.R
\name{ExactZonoVol}
\alias{ExactZonoVol}
\title{Compute the exact volume of a zonotope.}
\usage{
ExactZonoVol(Matrix)
}
\arguments{
\item{Matrix}{The \eqn{d \times m} matrix that containes the segments that define the zonotope.}
}
\value{
The exact volume of the zonotope
}
\description{
Given the \eqn{d \times m} matrix that containes the m segments that define the d-dimensional zonotope, this function computes the sum of the determinants of all the \eqn{d \times d} submatrices.
}
\examples{

# compute the exact volume of a 5-dimensional zonotope defined as the Minkowski sum of 10 segments
zonotope = GenZonotope(5, 10)
vol = ExactZonoVol(zonotope)
}
