% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenProdSimplex.R
\name{GenProdSimplex}
\alias{GenProdSimplex}
\title{Generator function for product of simplices.}
\usage{
GenProdSimplex(DimGen, repr = "H")
}
\arguments{
\item{Dimension}{The dimension of the simplex.}
}
\value{
A polytope defined as the product of two unit simplices in H-representation.The retutn value is a list with two elements: the "matrix" containing a \eqn{2d+1\times 2d} matrix A and the "vector" containing a 2d+1-dimensional vector b, s.t. \eqn{Ax\leq b}.
}
\description{
This function can be used to generate a 2d-dimensional polytope that is defined as the product of two d-dimensional unit simplices.
}
\examples{
# generate a product of two 5-dimensional simplices.
PolyList = GenProdSimplex(5)
}
