% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenProdSimplex.R
\name{GenProdSimplex}
\alias{GenProdSimplex}
\title{Generator function for product of simplices}
\usage{
GenProdSimplex(dimension)
}
\arguments{
\item{dimension}{The dimension of the simplices.}
}
\value{
A polytope class representing the product of the two \eqn{d}-dimensional unit simplices in H-representation.
}
\description{
This function can be used to generate a \eqn{2d}-dimensional polytope that is defined as the product of two \eqn{d}-dimensional unit simplices in H-representation.
}
\examples{
# generate a product of two 5-dimensional simplices.
P = GenProdSimplex(5)
}
