% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenSkinnyCube.R
\name{GenSkinnyCube}
\alias{GenSkinnyCube}
\title{Generator function for skinny hypercubes}
\usage{
GenSkinnyCube(dimension)
}
\arguments{
\item{dimension}{The dimension of the skinny hypercube.}
}
\value{
A polytope class representing the \eqn{d}-dimensional skinny hypercube in H-representation.
}
\description{
This function can be used to generate a \eqn{d}-dimensional skinny hypercube in H-representation.
}
\examples{
# generate a 10-dimensional skinny hypercube.
P = GenSkinnyCube(10)
}
