% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{geweke}
\alias{geweke}
\title{Geweke's MCMC diagnostic}
\usage{
geweke(samples, frac_first = NULL, frac_last = NULL)
}
\arguments{
\item{samples}{A matrix that contans column-wise the sampled points from a geometric random walk.}

\item{frac_first}{Optional. The portion of the first in order points in matrix samples.}

\item{frac_last}{Optional. The portion of the last in order points in matrix samples.}
}
\value{
A boolean to denote if the result of Geweke diagnostic: (i)  false if the null hypothesis is rejected, (ii) true if the null hypothesis is not rejected.
}
\description{
Geweke's MCMC diagnostic
}
\references{
\cite{Geweke, J.,
\dQuote{Evaluating the accuracy of sampling-based approaches to the calculation of posterior moments,} \emph{ In Bayesian Statistics 4. Proceedings of the Fourth Valencia International Meeting,} 1992.}
}
