% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{ineToMatrix}
\alias{ineToMatrix}
\title{functiion to get a ine file and return matrix A in ine format for VolEsti()}
\usage{
ineToMatrix(P)
}
\arguments{
\item{read.cs('path/to/file.ine')}{the ine file desrcibing the H-polytope}
}
\value{
The numerical matrix in ine format of \code{read.cs('path/to/file.ine')}
}
\description{
functiion to get a ine file and return matrix A in ine format for VolEsti()
}
\examples{
ineToMatrix(read.cs('path/to/data/cube40.ine'))
}
