% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{load_sdpa_format_file}
\alias{load_sdpa_format_file}
\title{An internal Rccp function to read a SDPA format file}
\usage{
load_sdpa_format_file(input_file = NULL)
}
\arguments{
\item{input_file}{Name of the input file}
}
\value{
A list with two named items: an item "matrices" which is a list of the matrices and an vector "objFunction"
}
\description{
An internal Rccp function to read a SDPA format file
}
\keyword{internal}
