% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{loadSdpaFormatFile}
\alias{loadSdpaFormatFile}
\title{Read a SDPA format file}
\usage{
loadSdpaFormatFile(inputFile = NULL)
}
\arguments{
\item{inputFile}{Name of the input file}
}
\value{
A list with two named items: an item "matrices" which is a list of the matrices and an vector "objFunction"
}
\description{
Read a SDPA format file
}
\examples{
path = system.file('extdata', package = 'volesti')
l = loadSdpaFormatFile(paste0(path,'/sdpa_n2m3.txt'))
}
