% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifyMat.R
\name{modifyMat}
\alias{modifyMat}
\title{takes a numerical matrix in ine format and return numerical matrix A and vector b: Ax<=b}
\usage{
modifyMat(A)
}
\arguments{
\item{A}{the numerical matrix in ine format of the H-polytope}
}
\value{
numerical matrix A and vector b: Ax<=b
}
\description{
takes a numerical matrix in ine format and return numerical matrix A and vector b: Ax<=b
}
\examples{
modifyMat(A)
}
