% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_lipschitz_constant.R
\name{pinvweibull_with_loc}
\alias{pinvweibull_with_loc}
\title{Inverse weibull distribution CDF with location parameter}
\usage{
pinvweibull_with_loc(q, k, lambda, theta)
}
\arguments{
\item{k}{The shape parameter}

\item{lambda}{The scale parameter}

\item{theta}{The location parameter}

\item{x}{The argument of the CDF}
}
\value{
The value of the CDF of an Inverse Weibull distribution with parameters k, lambda, theta evaluated at x
}
\description{
Inverse weibull distribution CDF with location parameter
}
