% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rand_rotate.R
\name{rand_rotate}
\alias{rand_rotate}
\title{Apply a random rotation to a convex polytope (H-polytope, V-polytope or a zonotope)}
\usage{
rand_rotate(P)
}
\arguments{
\item{P}{A convex polytope. It is an object from class (a) Hpolytope or (b) Vpolytope or (c) Zonotope.}
}
\value{
A random rotation of the polytope that is given as an input. The return class is the same as the input class.
}
\description{
Given a convex H or V polytope or a zonotope as input this function applies a random rotation.
}
\examples{
# rotate a H-polytope (2d unit simplex)
P = GenSimplex(2,'H')
P = rand_rotate(P)

# rotate a V-polytope (3d cube)
P = GenCube(3, 'V')
P = rand_rotate(P)

# rotate a 5-dimensional zonotope defined by the Minkowski sum of 15 segments
Z = GenZonotope(3,6)
Z = rand_rotate(Z)
}
