% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/round_polytope.R
\name{round_polytope}
\alias{round_polytope}
\title{Apply rounding to a convex polytope.}
\usage{
round_polytope(Inputs)
}
\arguments{
\item{list("argument"=value)}{A list that includes parameters for the rounding.}

\item{path}{The path to an ine (H-polytope) or ext (V-polytope, zonotope) file that describes the polytope. If path is given then "matrix" and "vector" inputs are not needed.}

\item{matrix}{The \eqn{m\times d} matrix A of the H polytope or the \eqn{m\times d} matrix that containes all the \eqn{m} d-dimensional vertices of a V polytope row-wise or a \eqn{m\times d} matrix that containes all the \eqn{m} d-dimensional segments that define a zonotope row-wise. If the matrix is in ine format, for H-polytopes only (see examples), then the "vector" input is not needed.}

\item{vector}{Only for H-polytopes. The m-dimensional vector b that containes the constants of the m facets, s.t.: \eqn{Ax\leq b}.}

\item{Vpoly}{A boolean parameter, has to be true when a V-polytope is given as input. Default value is false.}

\item{Zonotope}{A boolean parameter, has to be true when a zonotope is given as input. Default value is false.}

\item{walk_length}{Optional. The number of the steps for the random walk, default is \eqn{\lfloor 10+d/10\rfloor}.}

\item{ball_walk}{Optional. Boolean parameter to use ball walk, only for CG algorithm .Default value is false.}

\item{delta}{Optional. The radius for the ball walk.}

\item{coordinate}{Optional. A boolean parameter for the hit-and-run. True for Coordinate Directions HnR, false for Random Directions HnR. Default value is true.}

\item{verbose}{Optional. A boolean parameter for printing. Default is false.}
}
\value{
Is a list that contains elements to describe the rounded polytope, i.e. "matrix" and "vector" for H-polytopes and just "matrix" for V-polytopes and zonotopes, containing the verices or segments row-wise. For both representations the list contains element "round_value" which is the determinant of the square matrix of the linear transformation that was applied on the polytope that is given as input.
}
\description{
Given a convex H or V polytope or a zonotope as input this function computes a rounding based on minimum volume enclosing ellipsoid of a pointset.
}
\examples{
# rotate a H-polytope (2d unit simplex)
A = matrix(c(-1,0,0,-1,1,1), ncol=2, nrow=3, byrow=TRUE)
b = c(0,0,1)
listHpoly = round_polytope(list("matrix"=A, "vector"=b))

# rotate a V-polytope (3d cube) using Random Directions HnR
V = matrix(c(-1,1,-1,-1,-1,1,-1,1,1,-1,-1,-1,1,1,-1,1,-1,1,1,1,1,1,-1,-1), ncol=3, nrow=8, byrow=TRUE)
ListVpoly = round_polytope(list("matrix"=V, "Vpoly"=TRUE, "coordinate"=FALSE))

# rotate a 10-dimensional zonotope defined by the Minkowski sum of 20 segments
Zono = GenZonotope(10,20)
ListZono = round_polytope(list("matrix"=Zono, "Zonotope"=TRUE))
}
