% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rounding}
\alias{rounding}
\title{Internal rcpp function for the rounding of a convex polytope}
\usage{
rounding(P, random_walk = NULL, walk_length = NULL,
  parameters = NULL)
}
\arguments{
\item{P}{A convex polytope (H- or V-representation or zonotope).}

\item{random_walk}{Optional. A string that declares the random walk.}

\item{walk_length}{Optional. The number of the steps for the random walk.}

\item{parameters}{Optional. A list for the parameters of the methods:}
}
\value{
A numerical matrix that describes the rounded polytope and contains the round value.
}
\description{
Internal rcpp function for the rounding of a convex polytope
}
\section{warning}{

Do not use this function.
}

