% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rounding}
\alias{rounding}
\title{Internal rcpp function for the rounding of a convex polytope}
\usage{
rounding(P)
}
\arguments{
\item{P}{A convex polytope (H- or V-representation or zonotope).}
}
\value{
A numerical matrix that describes the rounded polytope and contains the round value.
}
\description{
Internal rcpp function for the rounding of a convex polytope
}
\section{warning}{

Do not use this function.
}

