% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rounding}
\alias{rounding}
\title{Internal rcpp function for the rounding of a convex polytope}
\usage{
rounding(P, method = NULL, seed = NULL)
}
\arguments{
\item{P}{A convex polytope (H- or V-representation or zonotope).}

\item{method}{Optional. The method to use for rounding, a) \code{'min_ellipsoid'} for the method based on mimimmum volume enclosing ellipsoid of a uniform sample from P, b) \code{'max_ellipsoid'} for the method based on maximum volume enclosed ellipsoid in P, (c) \code{'isotropy'} for the method based on isotropy. The default method is \code{'min_ellipsoid'} for all the representations.}

\item{seed}{Optional. A fixed seed for the number generator.}
}
\value{
A numerical matrix that describes the rounded polytope, a numerical matrix of the inverse linear transofmation that is applied on the input polytope, the numerical vector the the input polytope is shifted and the determinant of the matrix of the linear transformation that is applied on the input polytope.
}
\description{
Internal rcpp function for the rounding of a convex polytope
}
\keyword{internal}
