% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_points.R
\name{sample_points}
\alias{sample_points}
\title{Sample points from a convex Polytope}
\usage{
sample_points(Inputs)
}
\arguments{
\item{list("argument"=value)}{A list that includes parameters for the chosen target distribution and the random walk algorithm.}

\item{path}{The path to an ine or ext file that describes the H or V polytope respectively. If path is given then "matrix" and "vector" inputs are not needed.}

\item{matrix}{The matrix A of a H-polytope or the matrix V that contains all the vertices of a V polytope row-wise. If it is in ine format, only for H-polytopes, then the input "vector" is not needed.}

\item{vector}{Only for H-polytopes. The d-dimensional vector b that containes the constants of the facets.}

\item{walk_length}{Optional. The number of the steps for the random walk, default is \eqn{\lfloor 10+d/10\rfloor}.}

\item{internal_point}{Optional. A d-dimensional vector that containes an internal point of the polytope.}

\item{gaussian}{Optional. A boolean parameter to sample with gaussian target distribution. Default value is false.}

\item{variance}{Optional. The variance for the spherical gaussian. Default value is \eqn{1}.}

\item{N}{The number of points that the function is going to sample from the convex polytope. Default value is \eqn{100}.}

\item{ball_walk}{Optional. Boolean parameter to use ball walk for the sampling. Default value is false.}

\item{delta}{Optional. The radius for the ball walk.}

\item{verbose}{Optional. A boolean parameter for printing. Default is false.}

\item{vpoly}{A boolean parameter, has to be true when a V-polytope is given as input. Default value is false.}

\item{coordinate}{Optional. A boolean parameter for the hit-and-run. True for Coordinate Directions HnR, false for Random Directions HnR. Default value is true.}
}
\value{
A \eqn{d\times N} matrix that contains, column-wise, the sampled points from the convex polytope.
}
\description{
Sample N points from a H or a V-polytope with uniform or spherical gaussian target distribution.
}
\examples{
#uniform distribution from a 3d cube described by a set of vertices
V = matrix(c(-1,1,-1,-1,-1,1,-1,1,1,-1,-1,-1,1,1,-1,1,-1,1,1,1,1,1,-1,-1), ncol=3, nrow=8, byrow=TRUE)
points = sample_points(list("matrix"=V, "Vpoly"=TRUE, "N"=1000))

#gaussian distribution from a 2d unit simplex in H-representation with variance = 2
A = matrix(c(-1,0,0,-1,1,1), ncol=2, nrow=3, byrow=TRUE)
b = c(0,0,1)
vol = sample_points(list("matrix"=A, "vector"=b, "gaussian"=TRUE, "variance"=2))
}
