% Generated by hand
\name{vol_R}
\alias{vol_R}
\title{The main Rcpp function}
\usage{
vol_R(A, walk_len, e, InnerVec, CG, win_len, N, C, ratio, frac, ball_walk,
 delta, Vpoly, Zono, exact_zono, gen_only, Vpoly_gen, kind_gen, dim_gen,
 m_gen, round_only, rotate_only, ball_only, sample_only, sam_simplex,
 sam_can_simplex, sam_arb_simplex, sam_ball, sam_sphere, numpoints,
 variance, construct_copula, hyplane1, hyplane2, num_slices, sliceSimplex,
 coord, rounding, verbose)
}
\arguments{
\item{A}{A matrix in ine format (for H-polytopes) or in ext format (for V-polytopes or zonotopes).}

\item{walk_len}{The number of the steps for the random walk.}

\item{e}{The goal for the approximation error.}

\item{InnerVec}{A \eqn{d+1} vector that containes an inner ball. Or a \eqn{d} vector for a internal point when sample_point calls vol_R, or a vector with differnet size when there is not an input for this feature.}

\item{CG}{A boolean flag to use CoolingGaussian algorithm or to sample frm spherical gaussian when sample_only flag is true.}

\item{win_len}{The size of the window for the ratios' approximation in CG algorithm.}

\item{N}{The number of points we sample in each step of schedule annealing in CG algorithm.}

\item{C}{A constant for the lower bound of \eqn{variance/mean^2} in schedule annealing of CG algorithm.}

\item{ratio}{Parameter of schedule annealing of CG algorithm, larger ratio means larger steps in schedule annealing.}

\item{frac}{The fraction of the total error to spend in the first gaussian in CG algorithm.}

\item{ball_walk}{Boolean flag to use ball walk method for the random walk.}

\item{delta}{The radius for the ball walk.}

\item{Vpoly}{A boolean flag to declare if the input is a V-polytope.}

\item{Zono}{A boolean flag to declare if the input is a V-polytope.}

\item{exact_zono}{A boolean flag, it has to be true when the function is called to compute the exact volume of a zonotope.}

\item{gen_only}{A boolean flag, it has to be true when the function is called to generate a polytope.}

\item{Vpoly_gen}{A boolean flag, it has to be true when gen_only is true and the requested representation is the V-representation.}

\item{kind_gen}{An integer parameter to declare the kind of the polytope when the gen_only flag is true.}

\item{dim_gen}{An integer parameter to declare the dimension when the gen_only or sam_simplex, sam_can_simplex, sam_arb_simplex flags are true.}

\item{m_gen}{An integer parameter to declare the number of segments that generate a zonotope when gen_only flag is true.}

\item{round_only}{A boolean flag, it has to be true when the function is called only for rounding.}

\item{rotate_only}{A boolean flag, it has to be true when the function is called only for a random rotation of a convex polytope.}

\item{ball_only}{A boolean flag, it has to be true when the function is called only for the computation of an inner ball.}

\item{sample_only}{A boolean flag, it has to be true when the function is called only to sample from a convex polytope.}

\item{sam_simplex}{A boolean flag, it has to be true when the function is called only to sample uniformly from the unit simplex.}

\item{sam_can_simplex}{A boolean flag, it has to be true when the function is called only to sample uniformly from the canonical simplex.}

\item{sam_arb_simplex}{A boolean flag, it has to be true when the function is called only to sample uniformly from an arbitrary simplex in V-representation.}

\item{sam_ball}{A boolean flag, it has to be true when the function is called only to sample uniformly from a ball.}

\item{sam_sphere}{A boolean flag, it has to be true when the function is called only to sample uniformly from a sphere.}

\item{numpoints}{The number of points that the function is going to sample from the convex polytope when sample_only or sam_simplex, sam_can_simplex, sam_arb_simplex flags are true.}

\item{variance}{The variance for the spherical gaussian when sample_only flag is true.}

\item{construct_copula}{A boolean flag, it has to be true when the function is called only to construct a copula.}

\item{hyplane1}{A \eqn{d}-dimensional vector that containes the direction of a family of parallel hyperplanes when copula is constructed or a \eqn{d+1}-dimensional vector the containes a hyperplane when a slice of the unit simplex is computed.}

\item{hyplane2}{A \eqn{d}-dimensional vector that containes the direction of a family of parallel hyperplanes when copula is constructed.}

\item{num_slices}{An integer that declares the number of slices when a copula is constructed.}

\item{sliceSimplex}{A boolean flag, it has to be true when the percentage of the volume of a slice of the unit simplex is computed.}

\item{coord}{A boolean flag for the hit-and-run. True for Coordinate Directions HnR, false for Random Directions HnR.}

\item{rounding}{A boolean flag to activate the rounding option.}

\item{verbose}{A boolean flag for printing.}
}
\value{
A numerical matrix. For polytope generation, rotating and rounding is a  matrix in ine or ext format. For sampling is a \eqn{d\times N} matrix. For the volume approximation is \eqn{1\times 1} numerical matrix. 
}
\description{
All the R functions (except GenCubes, GenCross, GenProdSimplex, GenSimplex, Gen SkinnyCube and GenZonotope which call polytope_generate) call this Rcpp function. This is an internal function and we do not suggest to use it. 
}
\examples{
# see the package's functions
}
