% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/volume.R
\name{volume}
\alias{volume}
\title{The main R function for volume approximation of a convex Polytope (H-polytope, V-polytope or a zonotope)}
\usage{
volume(A, b, V, G, walk_length, error, InnerVec, CG, win_len, C, N, ratio, frac,
  ball_walk, delta, verbose, coordinate, rounding)
}
\arguments{
\item{A}{Only for H-polytopes. The \eqn{m\times d} matrix \eqn{A} that containes the directions of the \eqn{m} facets.}

\item{b}{Only for H-polytopes. The \eqn{m}-dimensional vector \eqn{b} that containes the constants of the \eqn{m} facets s.t.: \eqn{Ax\leq b}.}

\item{V}{Only for V-polytopes. The \eqn{m\times d} matrix V that containes row-wise the \eqn{m} \eqn{d}-dimensional vertices of the polytope.}

\item{G}{Only for zonotopes. The \eqn{m\times d} matrix G that containes row-wise the \eqn{m} \eqn{d}-dimensional segments that define a zonotope.}

\item{walk_length}{Optional. The number of the steps for the random walk. Default value is \eqn{\lfloor 10 + d/10\rfloor}.}

\item{error}{Optional. Declare the goal for the approximation error. Default value is \eqn{1} for SequenceOfBalls and \eqn{0.2} for CoolingGaussian.}

\item{InnerVec}{Optional. A \eqn{d+1} vector that containes an inner ball. The first \eqn{d} coordinates corresponds to the center and the last one to the radius of the ball. If it is not given then for H-polytopes the Chebychev ball is computed, for V-polytopes \eqn{d+1} vertices are picked randomly and the Chebychev ball of the defined simplex is computed. For a zonotope that is defined by the Minkowski sum of \eqn{m} segments we compute the maximal \eqn{r} s.t.: \eqn{re_i\in Z} for all \eqn{i=1,\dots ,d}, then the ball centered at the origin with radius \eqn{r/\sqrt{d}} is an internal ball.}

\item{CG}{Optional. A boolean parameter to use CoolingGaussian algorithm. Default value is false.}

\item{win_len}{Optional. The size of the window for the ratios' approximation in CG algorithm. Default value is \eqn{4 \cdot dimension^2 + 500}.}

\item{C}{Optional. A constant for the lower bound of \eqn{variance/mean^2} in schedule annealing of CG algorithm.}

\item{N}{optional. The number of points we sample in each step of schedule annealing in CG algorithm. Default value is \eqn{500C + dimension^2 / 2}.}

\item{ratio}{Optional. Parameter of schedule annealing of CG algorithm, larger ratio means larger steps in schedule annealing. Default value is \eqn{1 - 1/dimension}.}

\item{frac}{Optional. The fraction of the total error to spend in the first gaussian in CG algorithm. Default value is \eqn{0.1}.}

\item{ball_walk}{Optional. Boolean parameter to use ball walk. Default value is false.}

\item{delta}{Optional. The radius for the ball walk.}

\item{verbose}{Optional. A boolean parameter for printing. Default value is false.}

\item{coordinate}{Optional. A boolean parameter for the hit-and-run. True for Coordinate Directions HnR, false for Random Directions HnR. Default value is true.}

\item{rounding}{Optional. A boolean parameter to activate the rounding option. Default value is false.}
}
\value{
The approximation of the volume of a convex polytope.
}
\description{
For the volume approximation can be used two algorithms. Either SequenceOfBalls or CoolingGaussian. A H-polytope with \eqn{m} facets is described by a \eqn{m\times d} matrix \eqn{A} and a \eqn{m}-dimensional vector \eqn{b}, s.t.: \eqn{Ax\leq b}. A V-polytope is described as a set of \eqn{d}-dimensional points. A zonotope is desrcibed by the Minkowski sum of \eqn{d}-dimensional segments.
}
\examples{
# calling SOB algorithm for a H-polytope (2d unit simplex)
A = matrix(c(-1,0,0,-1,1,1), ncol=2, nrow=3, byrow=TRUE)
b = c(0,0,1)
vol = volume(A=A, b=b)

# calling CG algorithm for a V-polytope (3d cube)
Vmat = GenSimplex(2,'V')
vol = volume(V=Vmat, CG=TRUE)

# calling CG algorithm for a 5-dimensional zonotope defined as the Minkowski sum of 10 segments
zonotope = GenZonotope(2, 4)
vol = volume(G=zonotope, rounding=TRUE, CG=TRUE)
}
\references{
\cite{I.Z.Emiris and V. Fisikopoulos,
\dQuote{Practical polytope volume approximation,} \emph{ACM Trans. Math. Soft.,} 2014.},

\cite{B. Cousins and S. Vempala, \dQuote{A practical volume algorithm,} \emph{Springer-Verlag Berlin Heidelberg and The Mathematical Programming Society,} 2015.}
}
